Repeated fields problem when part of word
If later instances of a field form part of a work ([\w\s-_]+ etc) that word
is subsituted. Problem seems to stem from using the macro to move to the next
field and activate autocomplete. As a hack, add an entry to the cm-completion
file with just a space. It gets inserted instead, bypassing normal sublime
autocomplete behaviour and then you can type as expected. Results in an unwanted
double space though.